+++ src/main/java/cn/health/mapper/User_Height_WeightMapper.java
@@ -4,5 +4,5 @@ import cn.health.domain.User_Height_Weight;
 
 public interface User_Height_WeightMapper {
     void add(User_Height_Weight user_height_weight);
-    User_Height_Weight selectuserWHbyId(Long user_id);//查找用户最新的体重
+    User_Height_Weight selectuserWHbyId(Integer user_id);//查找用户最新的体重
 }
